🤖 refactor: 컴포넌트 폴더 구조 개선 및 import 경로 최적화
