+++ src/components/common/index.ts
@@ -1,0 +1,2 @@
+export { default as BottomNavbar } from './BottomNavbar';
+export { default as GoalTab } from './GoalTab';
