+++ src/components/timer/ResultModal.tsx
