+++ src/components/timer/TabSwitcher.tsx
