+++ src/components/timer/TimeDisplay.tsx
